webnn/avgpool2d.html

<!doctype html>
<meta charset=utf-8>
<title>test pooling operations</title>
<link rel="help" href="https://webmachinelearning.github.io/webnn/#api-mlgraphbuilder-pool2d">
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>
<script src="./webnn-polyfill.js"></script>
<div id=log></div>
<script type="module">
  'use strict';
  import * as utils from './utils.js';
  let builder;
  setup(() => {
    const context = navigator.ml.createContext();
    builder = new MLGraphBuilder(context);
  })

  test(() => {
    const x = builder.input('x', {type: 'float32', dimensions: [1, 3, 5, 5]});
    const y = builder.averagePool2d(x);
    const graph = builder.build({y});
    const inputs = {
      'x': new Float32Array([
        -1.1289884,  0.34016284,  0.497431,    2.1915932,   0.42038894,
        -0.18261199, -0.15769927, -0.26465914, 0.03877424,  0.39492005,
        -0.33410737, 0.74918455,  -1.3542547,  -0.0222946,  0.7094626,
        -0.09399617, 0.790736,    -0.75826526, 0.27656242,  0.46543223,
        -1.2342638,  1.1549494,   0.24823844,  0.75670505,  -1.7108902,
        -1.4767597,  -1.4969662,  -0.31936142, 0.5327554,   -0.06070877,
        0.31212643,  2.2274113,   1.2775147,   0.59886885,  -1.5765078,
        0.18522178,  0.22655599,  0.88869494,  0.38609484,  -0.05860576,
        -0.72732115, -0.0046324,  -1.3593693,  -0.6295078,  1.384531,
        0.06825881,  0.19907428,  0.20298219,  -0.8399954,  1.3583295,
        0.02117888,  -1.0636739,  -0.30460566, -0.92678875, -0.09120782,
        -0.88333017, -0.9641269,  0.6065926,   -0.5830042,  -0.81138134,
        1.3569402,   1.2891295,   0.2508177,   0.20211531,  0.8832168,
        -0.19886094, -0.61088,    0.682026,    -0.5253442,  1.5022339,
        1.0256356,   1.0642492,   -0.4169051,  -0.8740329,  1.1494869,
      ]),
    };
    const outputs = {y: new Float32Array(utils.sizeOfShape([1, 3, 1, 1]))};
    graph.compute(inputs, outputs);
    const expected = [
      0.07170040239999997,
      0.05194737240000002,
      0.07117922839999995,
    ];
    assert_array_approx_equals_ulp('averagePool2d', outputs.y, expected, 0);
  }, 'global averagePool2d default');
</script>
